# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api version 1",
        "description": "Laravel project API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "email": "contact@example.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api"
        }
    ],
    "paths": {
...
        "/v1/profile": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "профили юзера",
                "operationId": "29d38ca8c9484ccda8fafeb87a121131",
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Profile"
                ],
                "summary": "Добавить профиль",
                "operationId": "df580409ea8b40f4db429b0212e14cb7",
                "requestBody": {
                    "description": "",
                    "required": true,
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "phone",
                                    "title",
                                    "address",
                                    "inn"
                                ],
                                "properties": {
                                    "phone": {
                                        "description": "Номер мобильного телефона",
                                        "type": "string",
                                        "default": "+********"
                                    },
                                    "title": {
                                        "description": "Название компании",
                                        "type": "string",
                                        "default": "ООО Луна"
                                    },
                                    "address": {
                                        "description": "Адрес компании",
                                        "type": "string",
                                        "default": "город Москва, улица Ленина, дом 4"
                                    },
                                    "inn": {
                                        "description": "Инн компании (12 цифр)",
                                        "type": "string",
                                        "default": ""
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "created",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "fail"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
        "/v1/profile/{inn}": {
            "get": {
                "tags": [
                    "Profile"
                ],
                "summary": "профиль",
                "operationId": "0b72b42f0ec3555d2117c06171b780ef",
                "parameters": [
                    {
                        "name": "inn",
                        "in": "path",
                        "description": "inn компании"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "ok",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Profile"
                ],
                "summary": "Удалить профиль",
                "operationId": "53a748494ee88af13d2f9ea1b853f8ed",
                "parameters": [
                    {
                        "name": "inn",
                        "in": "path",
                        "description": "inn компании"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "deleted"
                    },
                    "404": {
                        "description": "not found"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            },
            "patch": {
                "tags": [
                    "Profile"
                ],
                "summary": "Обновить указанные поля",
                "operationId": "2053a066c2790cab49b575e45ef2c2a5",
                "parameters": [
                    {
                        "name": "inn",
                        "in": "path",
                        "description": "inn компании"
                    }
                ],
                "requestBody": {
                    "description": "Данные для обновления (только изменяемые поля)",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "inn": {
                                        "type": "string",
                                        "example": "111111111111"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "Москва"
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "АО Пятерочка"
                                    },
                                    "phone": {
                                        "type": "string",
                                        "example": "+70000000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "updated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Profile data updated"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "unauthorized"
                    },
                    "500": {
                        "description": "database error"
                    }
                },
                "security": [
                    {
                        "sanctumAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctumAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "Bearer"
            }
        }
    },
...
}